pull datetime agg function into utils
